# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ProcessingFailure {
  error: string;
}

export interface LearningCapabilities {
  ai: boolean;
  ocr: boolean;
  youtube: boolean;
  message: string;
  maxAnalysisCharacters: number;
  maxSegments: number;
}

export interface SourceSegment {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 18000
     */
  text: string;
  /** @minimum 1 */
  page?: number;
  /** @minimum 0 */
  startSeconds?: number;
  /** @minimum 0 */
  endSeconds?: number;
}

export interface ArticleInput {
  /** @maxLength 2000 */
  url: string;
  consent: true;
}

export interface ArticleResult {
  title: string;
  url: string;
  segments: SourceSegment[];
  warnings: string[];
}

export interface AnalysisInput {
  consent: true;
  /**
     * @minItems 1
     * @maxItems 20
     */
  segments: SourceSegment[];
}

export interface Grounding {
  /**
     * @minLength 1
     * @maxLength 100
     */
  segmentId: string;
  /**
     * @minLength 8
     * @maxLength 1000
     */
  quote: string;
}

export type ConceptProposalKind = typeof ConceptProposalKind[keyof typeof ConceptProposalKind];


export const ConceptProposalKind = {
  concept: 'concept',
  prerequisite: 'prerequisite',
  example: 'example',
} as const;

export interface ConceptProposal {
  /**
     * @minLength 1
     * @maxLength 100
     */
  key: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  description: string;
  kind: ConceptProposalKind;
  reference: Grounding;
}

export interface RelationProposal {
  fromKey: string;
  toKey: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  label: string;
  reference: Grounding;
}

export type QuestionProposalKind = typeof QuestionProposalKind[keyof typeof QuestionProposalKind];


export const QuestionProposalKind = {
  mcq: 'mcq',
  flashcard: 'flashcard',
  short: 'short',
} as const;

export interface QuestionProposal {
  conceptKey: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  prompt: string;
  kind: QuestionProposalKind;
  /**
     * @maxItems 6
     * @items.minLength 1
     * @items.maxLength 1000
     */
  choices: string[];
  /**
     * @minimum 0
     * @maximum 5
     */
  correctChoice: number;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  answer: string;
  /** @maxLength 2000 */
  rubric: string;
  reference: Grounding;
}

export interface AnalysisResult {
  /** @maxItems 15 */
  concepts: ConceptProposal[];
  /** @maxItems 20 */
  relations: RelationProposal[];
  /** @maxItems 20 */
  questions: QuestionProposal[];
  warnings: string[];
}

export interface PageImageInput {
  /**
     * @maxLength 3000000
     * @pattern ^data:image/(png|jpeg);base64,
     */
  image: string;
  consent: true;
}

export interface PageTextResult {
  /** @maxLength 18000 */
  text: string;
  warnings: string[];
}

export interface HealthStatus {
  status: string;
}

/**
 * Processing unavailable, invalid input, or rate limit
 */
export type ProcessingErrorResponse = ProcessingFailure;