# lib/api-zod/src/generated/types/analysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SourceSegment } from './sourceSegment';

export interface AnalysisInput {
  consent: true;
  /**
     * @minItems 1
     * @maxItems 20
     */
  segments: SourceSegment[];
}