# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConceptProposal } from './conceptProposal';
import type { QuestionProposal } from './questionProposal';
import type { RelationProposal } from './relationProposal';

export interface AnalysisResult {
  /** @maxItems 15 */
  concepts: ConceptProposal[];
  /** @maxItems 20 */
  relations: RelationProposal[];
  /** @maxItems 20 */
  questions: QuestionProposal[];
  warnings: string[];
}