# lib/api-zod/src/generated/types/articleInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleInput {
  /** @maxLength 2000 */
  url: string;
  consent: true;
}