# lib/api-zod/src/generated/types/articleResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SourceSegment } from './sourceSegment';

export interface ArticleResult {
  title: string;
  url: string;
  segments: SourceSegment[];
  warnings: string[];
}