# lib/api-zod/src/generated/types/conceptProposal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConceptProposalKind } from './conceptProposalKind';
import type { Grounding } from './grounding';

export interface ConceptProposal {
  /**
     * @minLength 1
     * @maxLength 100
     */
  key: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  description: string;
  kind: ConceptProposalKind;
  reference: Grounding;
}