# lib/api-zod/src/generated/types/grounding.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Grounding {
  /**
     * @minLength 1
     * @maxLength 100
     */
  segmentId: string;
  /**
     * @minLength 8
     * @maxLength 1000
     */
  quote: string;
}