# lib/api-zod/src/generated/types/pageTextResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PageTextResult {
  /** @maxLength 18000 */
  text: string;
  warnings: string[];
}