# lib/api-zod/src/generated/types/questionProposal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Grounding } from './grounding';
import type { QuestionProposalKind } from './questionProposalKind';

export interface QuestionProposal {
  conceptKey: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  prompt: string;
  kind: QuestionProposalKind;
  /**
     * @maxItems 6
     * @items.minLength 1
     * @items.maxLength 1000
     */
  choices: string[];
  /**
     * @minimum 0
     * @maximum 5
     */
  correctChoice: number;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  answer: string;
  /** @maxLength 2000 */
  rubric: string;
  reference: Grounding;
}