# lib/api-zod/src/generated/types/relationProposal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Grounding } from './grounding';

export interface RelationProposal {
  fromKey: string;
  toKey: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  label: string;
  reference: Grounding;
}